# swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Micro Service Project - Coalition"
host: "coalition.com"
basePath: "/api/"
...
schemes:
  - "https"
  - "http"
paths:
...
  /servers/members:
    post:
      tags:
        - "Server/Member"
      summary : "Add user to server"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : body
          required: true
          schema:
            type: object
            properties: 
              serverID:
                type: integer
              userID:
                type: integer
      responses:
        "201":
          description: "successful operation"
        "404":
          description: "User or Server not found"
        "500":
          description: "Internal server error"
    get:
      tags:
        - "Server/Member"
      summary: "Get all members of all servers"
      produces:
        - "application/json"
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                server_id: 
                  type : integer
                user_id :
                  type : integer
                created_at :
                  type : string
        "500":
          description: "Internal server error"
    delete:
      tags:
        - "Server/Member"
      summary : "Delete member from server"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : body
          required: true
          schema:
            type: object
            properties: 
              serverID:
                type: integer
              userID:
                type: integer
      responses:
        "204":
          description: "successful operation"
        "404":
          description: "Server or User not found / User is not member"
        "500":
          description: "Internal server error"
...
  /servers/members/user/{userID} :
    get :
      tags:
      - "Server/Member"
      summary: "Get all servers of a user"
      produces:
        - "application/json"
      parameters:
      - in: path
        name: "userID"
        description: "ID of user"
        required: true
        type: integer
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                server_id :
                  type : integer
                created_at :
                  type : string
        "404":
          description: "User not found"
        "500":
          description: "Internal server error"
...
  /groups:
    post:
      tags:
        - "Group"
      summary : "Create group"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : "body"
          required: true
          schema:
            type: object
            properties: 
              ownerID:
                type: integer
              groupName:
                type: "string"
      responses:
        "201":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              groupID:
                type: integer
        "404":
          description: "User not found"
        "500":
          description: "Internal server error"
    get:
      tags:
        - "Group"
      summary: "Get all groups"
      produces:
        - "application/json"
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                group_id: 
                  type : integer
                group_name:
                  type : string
                owner_id :
                  type : integer
                conv_id :
                  type : integer
                created_at :
                  type : string
                updated_at :
                  type: string
        "500":
          description: "Internal server error"
  /groups/{groupID}:
    parameters:
      - name: "groupID"
        in: path
        description: "ID of group"
        required: true
        type: integer
    get:
      tags:
        - "Group"
      summary: "Get a group"
      produces:
        - "application/json"
      responses:
        "200":
          description:  "successful operation"
          schema:
            type: object
            properties:
              server_name:
                type : string
              owner_id :
                type : integer
              conv_id:
                type: integer
              created_at :
                type : string
              updated_at :
                type: string
        "404":
          description:  "Group not found"
        "500":
          description: "Internal server error"
    put:
      tags:
        - "Group"
      summary: "Change group by ID"
      produces:
        - "application/json"
      parameters:
        - in: body
          name : "body"
          required: true
          schema:
            type: object
            properties: 
              userID:
                type: integer
              groupName:
                type: string
      responses:
        "200":
          description:  "successful operation"
        "403":
          description:  "Access forbidden"
        "404":
          description:  "Group or User not found"
        "500":
          description: "Internal server error"
    delete:
      tags:
        - "Group"
      summary: "Delete group by ID"
      produces:
        - "application/json"
      parameters:
        - in: body
          name : "body"
          required: true
          schema:
            type: object
            properties: 
              userID:
                type: integer
      responses:
        "204":
          description:  "successful operation"
        "403":
          description:  "Access forbidden"
        "404":
          description:  "Group or User not found"
        "500":
          description: "Internal server error"
  /groups/members:
    post:
      tags:
        - "Group/Member"
      summary : "Add user to group"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : body
          required: true
          schema:
            type: object
            properties: 
              groupID:
                type: integer
              userID:
                type: integer
      responses:
        "201":
          description: "successful operation"
        "404":
          description: "User or Group not found"
        "500":
          description: "Internal server error"
    get:
      tags:
        - "Group/Member"
      summary: "Get all members of all groups"
      produces:
        - "application/json"
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                group_id: 
                  type : integer
                user_id :
                  type : integer
                created_at :
                  type : string
        "500":
          description: "Internal server error"
    delete:
      tags:
        - "Group/Member"
      summary : "Delete member from group"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : body
          required: true
          schema:
            type: object
            properties: 
              groupID:
                type: integer
              userID:
                type: integer
      responses:
        "204":
          description: "successful operation"
        "404":
          description: "Group or User not found / User is not member"
        "500":
          description: "Internal server error"
  /groups/members/group/{groupID} :
    get :
      tags:
        - "Group/Member"
      summary: "Get all members of a group"
      produces:
        - "application/json"
      parameters:
        - name: "groupID"
          in: path
          description: "ID of group"
          required: true
          type: integer
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                user_id :
                  type : integer
                created_at :
                  type : string
        "404":
          description: "Group not found"
        "500":
          description: "Internal server error"
  /groups/members/user/{userID} :
    get :
      tags:
        - "Group/Member"
      summary: "Get all groups of a user"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: "userID"
          description: "ID of user"
          required: true
          type: integer
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                group_id :
                  type : integer
                created_at :
                  type : string
        "404":
          description: "User not found"
        "500":
          description: "Internal server error"
...